backend/pyproject.toml
[project]
name = "app"
description = "API voor bureaublad"
readme = "README.md"
license = "EUPL-1.2"
requires-python = ">=3.13"
dependencies = [
    "caldav>=1.4.0",
    "fastapi[standard]>=0.115.11",
    "openai>=1.66.5",
    "pydantic-settings>=2.8.1",
    "python-jose>=3.4.0",
    "python-ulid>=2.8.0",
]
dynamic = ["version"]

[project.urls]
repository = "https://github.com/MinBZK/bureaublad"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pyright>=1.1.406",
    "ruff>=0.13.3",
]


[tool.uv]
package = true

[tool.ruff]
line-length = 120
target-version = "py313"
src = ["app"]

[tool.ruff.lint]
select = ["I", "SIM", "B", "UP", "F", "E", "S", "C90", "DTZ", "LOG", "PIE", "PT", "ERA", "W", "C", "TRY", "RUF", "ANN"]
fixable = ["ALL"]
task-tags = ["TODO"]
ignore = ["TRY003", "C901"]

[tool.ruff.lint.mccabe]
max-complexity = 8

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "strict"
reportMissingImports = true
reportMissingTypeStubs = true
reportUnnecessaryIsInstance = false
exclude = [
    ".venv",
]
